+++ tests/test_symbols_tree_shitter.py
@@ -4,7 +4,7 @@
 # ScanCode is a trademark of nexB Inc.
 # SPDX-License-Identifier: Apache-2.0
 # See http://www.apache.org/licenses/LICENSE-2.0 for the license text.
-# See https://github.com/nexB/source-inspector for support or download.
+# See https://github.com/aboutcode-org/source-inspector for support or download.
 # See https://aboutcode.org for more information about nexB OSS projects.
 #
 
@@ -38,12 +38,14 @@ def check_json(expected, results, regen=REGEN_TEST_FIXTURES):
 
 class TestPygmentsSymbolScannerPlugin(FileBasedTesting):
 
-    test_data_dir = os.path.join(os.path.dirname(__file__), "data/symbols_tree_sitter")
+    test_data_dir = os.path.join(os.path.dirname(
+        __file__), "data/symbols_tree_sitter")
 
     def test_symbols_scanner_basic_cli_cpp(self):
         test_file = self.get_test_loc("test3.cpp")
         result_file = self.get_temp_file("json")
-        args = ["--treesitter-symbol-and-string", test_file, "--json-pp", result_file]
+        args = ["--treesitter-symbol-and-string",
+                test_file, "--json-pp", result_file]
         run_scan_click(args)
 
         expected_loc = self.get_test_loc("test3.cpp-expected.json")
@@ -52,7 +54,8 @@ class TestPygmentsSymbolScannerPlugin(FileBasedTesting):
     def test_symbols_scanner_long_cli(self):
         test_file = self.get_test_loc("if_ath.c")
         result_file = self.get_temp_file("json")
-        args = ["--treesitter-symbol-and-string", test_file, "--json-pp", result_file]
+        args = ["--treesitter-symbol-and-string",
+                test_file, "--json-pp", result_file]
         run_scan_click(args)
 
         expected_loc = self.get_test_loc("if_ath.c-expected.json")