+++ CHANGELOG.rst
@@ -1,10 +1,22 @@
 Changelog
 =========
 
+v0.5.2
+------
+
+- Update link references of ownership from nexB to aboutcode-org
+
+
+v0.5.1
+------
+
+- Update README with Pygments and Tree-Sitter example by @keshav-space in https://github.com/aboutcode-org/source-inspector/issues/22
+- Pin tree-sitter dependencies by @keshav-space in https://github.com/aboutcode-org/source-inspector/issues/23
+
 v0.5.0
 ------
 
-- Add support to collect strings, comments and symbols from source using Pygments 
+- Add support to collect strings, comments and symbols from source using Pygments
 - Add support to collect strings and symbols from source using tree-sitter